Fix old GCC error

+++ userspace/libsanalyzer/procfs_parser.cpp
@@ -284,7 +284,7 @@ double sinsp_procfs_parser::get_process_cpu_load_and_mem(uint64_t pid, uint64_t*
 	char line[512];
 	char tmps[32];
 	double res = -1;
-	string path = string("/proc/") + to_string(pid) + "/stat";
+	string path = string("/proc/") + to_string((long long unsigned int) pid) + "/stat";
 	uint64_t tval, val1, val2, val3, val4;
 
 	if(!m_is_live_capture)
